feat: Moved book carousel to differrent component
