+++ client/app/dashboard/dashboard.style.css
@@ -34,9 +34,16 @@
 	box-shadow: rgba(50, 50, 93, 0.25) 0px 6px 12px -2px, rgba(0, 0, 0, 0.3) 0px 3px 7px -3px;
 }
 
+.carousel-title {
+	padding: 1rem 0;
+	text-align: center;
+	width: 100%;
+}
+
 .book-list-item {
 	border: 1px solid brown;
 	padding: 5px 1rem;
+	margin: 0 2rem;
 }
 
 .book-list-item:hover{
