+++ client/app/layout.tsx
@@ -1,5 +1,6 @@
 import { MantineProvider, createTheme } from '@mantine/core';
 import { Notifications } from '@mantine/notifications';
+import '@mantine/carousel/styles.css';
 
 export default function RootLayout({children,}: {children: React.ReactNode}) {
 	const theme = createTheme({
